Minor restructuring of binding code
