+++ bundles/org.openhab.binding.omnilink/src/main/java/org/openhab/binding/omnilink/internal/TemperatureFormat.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.omnilink.internal.handler;
+package org.openhab.binding.omnilink.internal;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
